re-organize some functionalities
databroker  is now: nsls2ptycho.databroker
ptycho      is now: nsls2ptycho.backend.ptycho
xcale      will be: nsls2ptycho.backend.xcale

open question: what should be kept in nsls2ptycho.core?
