+++ nsls2ptycho/__init__.py
@@ -1,4 +1,6 @@
+import nsls2ptycho.backend
 import nsls2ptycho.core
+import nsls2ptycho.databroker
 import nsls2ptycho.ui
 from ._version import __version__
