+++ nsls2ptycho/databroker/HXN_databroker.py
@@ -20,7 +20,7 @@ except FileNotFoundError:
 
 
 # ***************************** "Public API" *****************************
-# The following functions must exist in nsls2ptycho.core.databroker,
+# The following functions must exist in nsls2ptycho.databroker,
 # but the function signatures do not need to agree across modules
 # (obviously, it is impossible for all beamlines to have the same setup).
 #   - load_metadata
