+++ nsls2ptycho/databroker/__init__.py
@@ -4,7 +4,7 @@ import platform
 
 
 # ***************************** "Public API" *****************************
-# The following functions must exist in nsls2ptycho.core.databroker,
+# The following functions must exist in nsls2ptycho.databroker,
 # but the function signatures do not need to agree across modules
 # (obviously, it is impossible for all beamlines to have the same setup).
 #   - load_metadata
